pyproject.toml
[project]
name = "data-preprocess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2025.12.0",
    "hdbscan>=0.8.41",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pybaseball>=2.2.7",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "umap-learn>=0.5.11",
]
